pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "llm_factory_toolkit"
version = "0.8.0"
authors = [
  { name="Diego Carboni", email="carboni123@hotmail.com" },
]
description = "A toolkit for interacting with various Large Language Models"
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Intended Audience :: Developers",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

# Only core deps here – no LLM backends
dependencies = [
    "python-dotenv>=1.1",
    "pydantic>=2.11",
]

[project.urls]
"Homepage" = "https://github.com/carboni123/llm_factory_toolkit"
"Bug Tracker" = "https://github.com/carboni123/llm_factory_toolkit/issues"

[project.optional-dependencies]
# LLM backends – user can choose one or many
openai = [
    "openai>=2.7.1",
]
google = [
    "google-genai>=1.53.0",
]

# “Convenience” extra to install all backends at once
llms = [
    "openai>=2.7.1",
    "google-genai>=1.53.0",
]

dev = [
   "pytest>=9.0.1",
   "pytest-asyncio>=1.3.0",
]
builtins = [
   "sympy>=1.12",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
    "asyncio: mark a test as running with asyncio support",
]
